Change rate-limiting to dynamic (based on headers)
Add different instance for the stub api
Fix CacheTest interference on other tests
Add @Ignore to long tests
Use SecretFile in UseageTest so it runs
Update Readme

+++ src/main/java/no/stelar7/api/l4j8/basic/ratelimiting/RateLimiter.java
@@ -30,7 +30,7 @@ public abstract class RateLimiter
     
     public abstract void acquire();
     
-    public abstract void updatePermitsPerX(Map<Integer, Long> data);
+    public abstract void updatePermitsPerX(Map<Integer, Integer> data);
     
     public Map<RateLimit, Instant> getFirstCallInTime()
     {
